pyproject.toml
[project]
name = "a-webhook-receiver"
version = "0.1.0"
description = "Webhook receiver prototype"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiosqlite>=0.20.0",
    "prometheus-client>=0.21.0",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/webhook_receiver"]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]  # FastAPI Depends() in defaults is intentional

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=webhook_receiver --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["src/webhook_receiver"]
omit = ["*/schema.sql"]
branch = true

[tool.coverage.report]
show_missing = true
